+++ src/main/java/cse213/final_project/Jannat/GeneralCitizen.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 
 public class GeneralCitizen extends User implements Serializable {
 
-    public GeneralCitizen(String address, LocalDate dob, String email, int id, String name, String password, String pN, String type) {
+    public GeneralCitizen(String address, LocalDate dob, String email, Long id, String name, String password, String pN, String type) {
         super(address, dob, email, id, name, password, pN, type);
     }
 }
