+++ src/main/java/cse213/final_project/Jannat/VerificationOfficer.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 import java.time.LocalDate;
 
 public class VerificationOfficer extends User  implements Serializable {
-    public VerificationOfficer(String address, LocalDate dob, String email, int id, String name, String password, String pN, String type) {
+    public VerificationOfficer(String address, LocalDate dob, String email, Long id, String name, String password, String pN, String type) {
         super(address, dob, email, id, name, password, pN, type);
     }
 }
